Added more exceptions and moved them to packages
