+++ data/linus.txt
@@ -1,7 +1,4 @@
 T;;0;;asg
 T;;1;;asfd sgd
 T;;1;;asfa
-T;;0;;ashd
-T;;0;;12312
-E;;0;;winning;;12/04/2020 1900;;31/03/2021 2000
 T;;0;;asdaf
