--- src/main/java/duke/EmptyMessegeException.java
@@ -1,6 +1,0 @@
-package duke;
-public class EmptyMessegeException extends DukeCeption {
-    public EmptyMessegeException(String error) {
-        super(error);
-    }
-}
