+++ src/main/java/duke/command/CommandUnmark.java
@@ -1,9 +1,9 @@
 package duke.command;
 
-import duke.Ui;
-import duke.tasks.Task;
 import duke.TaskList;
-import duke.DukeCeption;
+import duke.Ui;
+import duke.exceptions.DukeCeption;
+import duke.tasks.Task;
 
 public class CommandUnmark extends Command {
