+++ src/main/java/duke/exceptions/DukeCeption.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.exceptions;
 public class DukeCeption extends Exception {
     public DukeCeption(String error) {
         super(error);
