+++ src/main/java/duke/exceptions/EmptyDescriptionException.java
@@ -1,0 +1,6 @@
+package duke.exceptions;
+public class EmptyDescriptionException extends DukeCeption {
+    public EmptyDescriptionException(String error) {
+        super(error);
+    }
+}
