+++ src/main/java/duke/exceptions/IncorrectFormatException.java
@@ -1,0 +1,6 @@
+package duke.exceptions;
+public class IncorrectFormatException extends DukeCeption {
+    public IncorrectFormatException(String error) {
+        super(error);
+    }
+}
