+++ src/main/java/duke/exceptions/NumberOutOfBoundsException.java
@@ -1,0 +1,6 @@
+package duke.exceptions;
+public class NumberOutOfBoundsException extends DukeCeption {
+    public NumberOutOfBoundsException(String error) {
+        super(error);
+    }
+}
